Add years to Discover carousels
