+++ Managers/Library/LMPinnedItems.swift
@@ -313,7 +313,7 @@ extension LibraryManager {
             } else if let album = entity as? AlbumEntity {
                 await self.pinAlbumEntity(album)
             } else if let category = entity as? CategoryEntity {
-                // Discover genre/decade tiles have no PinnedItem to route through.
+                // Discover category tiles have no PinnedItem to route through.
                 await self.pinLibraryItem(filterType: category.filterType, filterValue: category.name)
             }
         }
